ISPRGG1 phantom odyssey winforms rpg
Phantom Odyssey is a horror-based, 8-bit RPG game with elements of an adventure, choice-based narrative. We aimed to implement 2 endings and a mini-game by utilizing trigger event buttons; including but not limited to form Load and form Picture Boxes. 

We chose to develop Phantom Odyssey as we would like to experiment with RPG development. With both of us, Ms. Gabrielle Almirol and Ms. Andrea Ochotorena, working on the conceptualization, dialogues, and the design of the setting, we experienced how to implement 2 primary classes, Character and Phantom, and manipulate their attributes in the first minigame to pose an obstacle amidst the RPG narrative. 

The front-end and design development was implemented by Ms. Andrea Ochotorena: using Picture Boxes such as previous and next to trigger the narrative text and dialogues, we are quickly immersed into the Phantom Odyssey setting. The theme and color scheme was hugely influenced by multiple horror films, such as the Saw III. If-else statements were used to switch between previous and next dialogues; furthermore, we chose to implement multiple windows forms to change the setting and program it to link to each other. 

The back-end development was implemented by Ms. Gabrielle Almirol: utilizing the Character and Phantom classes and instantiating them on the Retaliate and Run Picture Boxes. The difficulty of the first minigame is implemented by setting a timer of 10 seconds wherein the user figures out how the Retaliate and Run buttons, wherein the attributes are coded into the Character and Phantom classes. By instantiating the Attack() method from the Phantom class, the hero’s health decreases as the player clicks on the Retaliate button; the Phantom draws nearer the more the health level decreases, indicating that the user is losing. On the contrary, the Run() method calls on the Run picture box increases the instantiated character’s health level until it reaches the health maximum. Once the timer reaches 0, the window closes and is redirected to the next window; the Game Over or Game Win forms, depending if the health level is equal to or greater than 0.